+++ lib/src/replay/image_extension.dart
@@ -1,0 +1,5 @@
+import 'dart:ui' as ui;
+
+extension ImageExtension on ui.Image {
+  bool get isValidSize => width > 0 && height > 0;
+}